refactor: EquipmentRentalSpecsResponse 레코드로 한 파일로 모음
