+++ src/main/java/com/girigiri/kwrental/rental/dto/response/EquipmentRentalSpecsResponse.java
@@ -1,18 +1,30 @@
 package com.girigiri.kwrental.rental.dto.response;
 
+import java.time.LocalDate;
 import java.util.List;
 
-import lombok.Getter;
+import lombok.Builder;
 
-@Getter
-public class EquipmentRentalSpecsResponse {
-    private List<EquipmentRentalSpecResponse> rentalSpecs;
+public record EquipmentRentalSpecsResponse(
+    List<EquipmentRentalSpecResponse> rentalSpecs) {
 
-    private EquipmentRentalSpecsResponse() {
-    }
+    @Builder
+    public record EquipmentRentalSpecResponse(
+        String status,
+        LocalDate acceptDate,
+        LocalDate returnDate,
+        String name,
+        String reason) {
 
-    private EquipmentRentalSpecsResponse(final List<EquipmentRentalSpecResponse> rentalSpecs) {
-        this.rentalSpecs = rentalSpecs;
+        public static EquipmentRentalSpecResponse from(final RentalSpecWithName rentalSpecWithName) {
+            return EquipmentRentalSpecResponse.builder()
+                .status(rentalSpecWithName.getStatus().isAbnormalReturned() ? "불량 반납" : "정상 반납")
+                .acceptDate(LocalDate.from(rentalSpecWithName.getAcceptDateTime()))
+                .returnDate(LocalDate.from(rentalSpecWithName.getReturnDateTime()))
+                .name(rentalSpecWithName.getName())
+                .reason(rentalSpecWithName.getStatus().name())
+                .build();
+        }
     }
 
     public static EquipmentRentalSpecsResponse from(final List<RentalSpecWithName> rentalSpecWithNames) {
